web/backend/requirements.txt
# FastAPI 后端依赖

# Web 框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# 深度学习
torch>=2.0.0
transformers>=4.37.0
accelerate>=0.27.0
peft>=0.8.0
bitsandbytes>=0.41.0

# 图像处理
pillow>=10.0.0
qwen-vl-utils>=0.0.1

# 工具
python-dotenv>=1.0.0
